Move api utility to httputil
