+++ pkg/util/httputil/request_test.go
@@ -1,4 +1,4 @@
-package api
+package httputil
 
 import (
 	"testing"
